# apps/backend/openapi/web-api.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RereviewRequest": {
        "properties": {
          "ticket_id": {
            "format": "uuid",
            "title": "Ticket Id",
            "type": "string"
          }
        },
        "required": [
          "ticket_id"
        ],
        "title": "RereviewRequest",
        "type": "object"
      },
      "ReviewJob": {
        "description": "Read-side projection of a `reviews` row for the ticket history API.",
        "properties": {
          "commit_sha_at_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Sha At Start"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "org_id": {
            "format": "uuid",
            "title": "Org Id",
            "type": "string"
          },
          "pr_id": {
            "format": "uuid",
            "title": "Pr Id",
            "type": "string"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "scope_kind": {
            "title": "Scope Kind",
            "type": "string"
          },
          "sequence_number": {
            "title": "Sequence Number",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trigger_reason": {
            "title": "Trigger Reason",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "org_id",
          "pr_id",
          "status",
          "trigger_reason",
          "destination",
          "scope_kind",
          "commit_sha_at_start",
          "sequence_number",
          "created_at",
          "updated_at"
        ],
        "title": "ReviewJob",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "yaaos",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/reviewer/cancel": {
      "post": {
        "description": "Cancel any non-terminal workflow_executions for this ticket.",
        "operationId": "cancel_jobs_api_reviewer_cancel_post",
        "parameters": [
          {
            "in": "query",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Slug"
            }
          },
          {
            "in": "cookie",
            "name": "yaaos_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yaaos Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Cancel Jobs Api Reviewer Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Jobs",
        "tags": [
          "reviewer"
        ]
      }
    },
    "/api/reviewer/findings/by-ticket/{ticket_id}": {
      "get": {
        "description": "List findings for the ticket's PR in the canonical schema.",
        "operationId": "findings_by_ticket_api_reviewer_findings_by_ticket__ticket_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_terminal",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Terminal",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Slug"
            }
          },
          {
            "in": "cookie",
            "name": "yaaos_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yaaos Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Findings By Ticket Api Reviewer Findings By Ticket  Ticket Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Findings By Ticket",
        "tags": [
          "reviewer"
        ]
      }
    },
    "/api/reviewer/jobs/by-ticket/{ticket_id}": {
      "get": {
        "description": "Per-ticket review history.",
        "operationId": "jobs_by_ticket_api_reviewer_jobs_by_ticket__ticket_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Slug"
            }
          },
          {
            "in": "cookie",
            "name": "yaaos_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yaaos Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewJob"
                  },
                  "title": "Response Jobs By Ticket Api Reviewer Jobs By Ticket  Ticket Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Jobs By Ticket",
        "tags": [
          "reviewer"
        ]
      }
    },
...
    "/api/reviewer/rereview": {
      "post": {
        "description": "Re-review a ticket — drives `pr_review_v1` via the workflow engine.",
        "operationId": "rereview_ticket_api_reviewer_rereview_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Org-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Slug"
            }
          },
          {
            "in": "cookie",
            "name": "yaaos_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yaaos Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RereviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Rereview Ticket Api Reviewer Rereview Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rereview Ticket",
        "tags": [
          "reviewer"
        ]
      }
    },
    "/api/reviewer/reviews/by-ticket/{ticket_id}": {
      "get": {
        "description": "Per-review timeline metadata.",
        "operationId": "reviews_by_ticket_api_reviewer_reviews_by_ticket__ticket_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Org-Slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Slug"
            }
          },
          {
            "in": "cookie",
            "name": "yaaos_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yaaos Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Reviews By Ticket Api Reviewer Reviews By Ticket  Ticket Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reviews By Ticket",
        "tags": [
          "reviewer"
        ]
      }
    },
...
  }
}